app/services/model_service.py
import torch
import torchvision.transforms as transforms
from PIL import Image
import io
from app.core.config import settings

class ModelService:
    def __init__(self):
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
            transforms.Normalize(mean=[0.485, 0.456, 0.406],
                              std=[0.229, 0.224, 0.225])
        ])
        self.model = self._load_model()

    def _load_model(self):
        # TODO: Implement model loading logic
        # This is a placeholder for the actual model loading
        return None

    async def predict(self, file):
        contents = await file.read()
        image = Image.open(io.BytesIO(contents))
        
        # Convert image to tensor
        image_tensor = self.transform(image).unsqueeze(0).to(self.device)
        
        # TODO: Implement actual prediction logic
        # This is a placeholder for the actual prediction
        return {
            "diagnosis": "Sample Diagnosis",
            "confidence": 0.95,
            "image_type": "X-Ray"
        } 